% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksvd-package.r
\docType{package}
\name{ksvd-package}
\alias{ksvd-package}
\title{ksvd}
\description{
TODO
}
\references{
Programming with Big Data in R Website: \url{https://pbdr.org/}
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
