% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksvd.r
\name{ksvd}
\alias{ksvd}
\title{Singular Value Decomposition}
\usage{
ksvd(x, nu = min(m, n), nv = min(m, n))
}
\arguments{
\item{x}{numeric distributed matrix}

\item{nu}{number of left singular vectors to return when calculating singular values.}

\item{nv}{number of right singular vectors to return when calculating singular values.}
}
\value{
TODO
}
\description{
TODO
}
\details{
TODO
}
