% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdgeqsvd.r
\name{rpdgeqsvd}
\alias{rpdgeqsvd}
\title{rpdgeqsvd}
\usage{
rpdgeqsvd(jobu, jobvt, eigtype, A, desca, descu, descvt)
}
\arguments{
\item{jobu}{TODO}

\item{jobvt}{TODO}

\item{eigtype}{TODO}

\item{A}{TODO}

\item{desca, descu, descvt}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\details{
For advanced users only.
}
